runanywhere-agent/pyproject.toml
[project]
name = "runanywhere-agent"
version = "1.0.0"
description = "AI coding agent for RunAnywhere SDK development - Build mobile apps with on-device AI"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = "Apache-2.0"
authors = [
    {name = "RunAnywhere AI", email = "founders@runanywhere.ai"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "aider-chat>=0.86.0",
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://www.runanywhere.ai"
Documentation = "https://github.com/RunanywhereAI/runanywhere-sdks"
Repository = "https://github.com/RunanywhereAI/runanywhere-sdks"

[project.scripts]
runanywhere-agent = "runanywhere_agent.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["runanywhere_agent*"]

[tool.setuptools.package-data]
runanywhere_agent = ["resources/*.md"]
